+++ client/src/app/app-routing.module.ts
@@ -1,7 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import {RegisterComponent} from "./security/register/register.component";
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'register', component: RegisterComponent }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
