+++ client/src/tests/functional/app.component.spec.ts
@@ -1,7 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AppComponent } from './app.component';
-import {HttpClientTestingModule} from "@angular/common/http/testing";
+import {AppComponent} from "../../app/app.component";
 
 describe('AppComponent', () => {
   beforeEach(async () => {
